Add accomodation create using APIs

+++ ssgso-project/src/main/java/com/project/ssgso/dto/AccomodationDto.java
@@ -5,8 +5,8 @@ public class AccomodationDto {
 	private String name;
 	private String info;
 	private String ac_img;
-	private int latitude;
-	private int longitude;
+	private Double latitude;
+	private Double longitude;
 	private String address;
 	
 	public AccomodationDto() {
@@ -44,23 +44,24 @@ public class AccomodationDto {
 	public void setAc_img(String ac_img) {
 		this.ac_img = ac_img;
 	}
-
-	public int getLatitude() {
+	
+	
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(int latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public int getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(int longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
-	
+
 	public String getAddress() {
 		return address;
 	}
